api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.16.4
google-genai==0.3.0
pydantic==2.5.0
numpy==1.24.3
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html